move the cade deletion to a new page, where you can choose wich sitemap you want to delete and fix some css stuff on other pages

+++ ajax/deleteCache.ajax.php
@@ -1,14 +1,18 @@
 <?php
+$sitemaps = $_POST['sitemaps'];
+
 $res = 0;
 $cache_folder = "../sitemap";
 $folder = opendir($cache_folder);
 
 while (false !== ($file = readdir($folder))){ // Reading each file of the folder
   $path = $cache_folder."/".$file;
-  
+
   // Check if file isn't a folder or .gitkeep
   if ($file != ".." AND $file != "." AND $file != ".gitkeep" AND !is_dir($file)){
-    unlink($path);
+    if (in_array($file, $sitemaps)){
+       unlink($path);
+    }
   }
 }
